# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HYDROPY SMS Gateway Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /devices:
    get:
      operationId: listDevices
      tags: [devices]
      summary: List all gateway devices
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [NEW, PENDING_CONNECTION, CONNECTED, ONLINE, BUSY, OFFLINE, BLOCKED]
      responses:
        "200":
          description: List of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      operationId: createDevice
      tags: [devices]
      summary: Register a new gateway device (generates QR token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceInput"
      responses:
        "201":
          description: Device created with QR data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceWithQr"
  /devices/{deviceId}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get a specific device
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDevice
      tags: [devices]
      summary: Remove a device
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /devices/{deviceId}/disconnect:
    post:
      operationId: disconnectDevice
      tags: [devices]
      summary: Force disconnect a device
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{deviceId}/qr:
    get:
      operationId: getDeviceQr
      tags: [devices]
      summary: Get or regenerate QR code for device pairing
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: QR code data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceQr"

  # ─── SIM Cards ──────────────────────────────────────────────────────────────
  /devices/{deviceId}/simcards:
    get:
      operationId: listSimCards
      tags: [simcards]
      summary: List SIM cards for a device
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SIM cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SimCard"

  # ─── Campaigns ──────────────────────────────────────────────────────────────
...
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages with optional filters
      parameters:
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: deviceId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [CREATED, QUEUED, ASSIGNED, SENDING, SENT, FAILED, DELIVERED]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageList"
  /messages/{messageId}:
    get:
      operationId: getMessage
      tags: [messages]
      summary: Get a specific message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ─── Auth ──────────────────────────────────────────────────────────────────
...
    DeviceInput:
      type: object
      required: [deviceName]
      properties:
        deviceName:
          type: string
          minLength: 1
    Device:
      type: object
      required: [id, deviceName, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        deviceName:
          type: string
        deviceToken:
          type: ["string", "null"]
        status:
          type: string
          enum: [NEW, PENDING_CONNECTION, CONNECTED, ONLINE, BUSY, OFFLINE, BLOCKED]
        androidVersion:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        appVersion:
          type: ["string", "null"]
        battery:
          type: ["integer", "null"]
        signal:
          type: ["integer", "null"]
        lastHeartbeat:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    DeviceWithQr:
      allOf:
        - $ref: "#/components/schemas/Device"
        - type: object
          required: [qrData]
          properties:
            qrData:
              type: string
    DeviceQr:
      type: object
      required: [deviceId, qrData, token]
      properties:
        deviceId:
          type: integer
        qrData:
          type: string
        token:
          type: string
    SimCard:
      type: object
      required: [id, deviceId, slot, status]
      properties:
        id:
          type: integer
        deviceId:
          type: integer
        slot:
          type: integer
        carrier:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [ACTIVE, INACTIVE, UNKNOWN]

    # ─── Campaigns ─────────────────────────────────────────────────────────────
...
    Message:
      type: object
      required: [id, recipient, message, status, createdAt]
      properties:
        id:
          type: integer
        campaignId:
          type: ["integer", "null"]
        deviceId:
          type: ["integer", "null"]
        recipient:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [CREATED, QUEUED, ASSIGNED, SENDING, SENT, FAILED, DELIVERED]
        failureReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        sentAt:
          type: ["string", "null"]
          format: date-time
    MessageList:
      type: object
      required: [messages, total]
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        total:
          type: integer