# lib/api-zod/src/generated/types/deviceInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * HYDROPY SMS Gateway Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DeviceInput {
  /** @minLength 1 */
  deviceName: string;
}