# lib/api-zod/src/generated/types/deviceWithQr.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * HYDROPY SMS Gateway Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Device } from './device';

export type DeviceWithQr = Device & {
  qrData: string;
};