# lib/api-zod/src/generated/types/simCard.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * HYDROPY SMS Gateway Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SimCardStatus } from './simCardStatus';

export interface SimCard {
  id: number;
  deviceId: number;
  slot: number;
  /** @nullable */
  carrier?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  status: SimCardStatus;
}